Update get_collapsed_candidates.py
Last version of get_collapsed_candidates.py that saves additional output files with projected and projection blocks coordinates in the same file. It generates four output files: the main output BED file with the final collapsed candidates and three additional files containing the coordinates of the original candidate blocks in the reference genome coordinate system with the projected blocks with coordinates based on the system of the reference genome, other haplotype, and the query genome.